requirements.txt
adrf==0.1.6
asgiref==3.8.1
async-property==0.2.2
click==8.1.7
Deprecated==1.2.14
Django==5.0.4
djangorestframework==3.15.1
googleapis-common-protos==1.63.0
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
importlib-metadata==7.0.0
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgiref @ git+https://github.com/xrmx/opentelemetry-instrumentation-asgiref.git@7b64ecfc08cb741a85b1ea24affc516090b262e7
opentelemetry-instrumentation-asyncio==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-test-utils==0.45b0
packaging==24.0
protobuf==4.25.3
psycopg==3.1.18
sqlparse==0.5.0
typing_extensions==4.11.0
uvicorn==0.29.0
wrapt==1.16.0
zipp==3.18.1
